Overview: Completed user profile generation functionality.
FIX: Logic in KL Divergence dip detection method so that it works correctly for a low number of topics.
ADD: Methods to support time factor calculations and processing.
ADD: Methods to support log file loading and processing into an LDA model.
ADD: A routine for generating a user profile.
MOD: Updated the log maker's output format to be more suitable for string processing.